python_engine/pyproject.toml
[project]
name = "fuel-hedging-platform"
version = "1.0.0"
description = "Aviation fuel hedging optimization platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.27.1",
    "sqlalchemy[asyncio]==2.0.27",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
    "pydantic[email]==2.6.1",
    "pydantic-settings==2.1.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "httpx==0.27.0",
    "structlog==24.1.0",
    "slowapi==0.1.9",
    "redis==5.0.1",
    "apscheduler==3.10.4",
    "pandas==2.2.0",
    "numpy==1.26.4",
    "scipy==1.12.0",
    "statsmodels==0.14.1",
    "scikit-learn==1.4.0",
    "xgboost==2.0.3",
    "tensorflow-cpu==2.15.0",
    "python-multipart==0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.0",
    "pytest-asyncio==0.23.4",
    "pytest-cov==4.1.0",
    "mypy==1.8.0",
    "ruff==0.2.1",
    "pre-commit==3.6.0",
    "httpx==0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --strict-markers --cov=app --cov-report=term-missing"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"
